refactor: move form related elements and logic to separate component
