+++ client/src/components/AddReviewModal/AddReviewModal.css
@@ -1,0 +1,5 @@
+.modalTitle {
+  background-color: #1976d2;
+  color: #fff;
+  margin-bottom: 10px;
+}
